Reimplement occupied cell in pathfinding
I knew this would be hard when reimplementing because of all the changes and wanting to make it as efficient and elegant as possible. In the end, I was having trouble figuring out when the next cell was empty and NOT the current entity since it stops to shoot and timing was thrown. I changed from a bool occupied table to simply a Node2D one so I could check if it was empty (null) AND not the same entity. Worked like a charm and was super easy!

Also added player invincibility so I could test the above! And will be good for other debugging. Or perhaps power ups?